Bug 5188 - [breakpoints] Breakpoints from one project, hit in another project
Added filter for breakpoints from resources outside of the scope of the
resources specified by the launch configuration. Additionally added
filter for types not accessible from the projects in scope (e.g
breakpoints in libraries not visible from the classpath of the running
VM).

To avoid possible performance issues with breakpoints without resource
data while updating breakpoints view, a type name cache is added which
remembers if a type was accessible or not in the current launch scope.

Note, that PDE like launch configurations which do not specify any
resources in the launch configuration, are not affected by this change.

Change-Id: I4253c22ae08bbbd478f4900301619b63e73bb2cc
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>