Bug 5188 - [breakpoints] Breakpoints from one project, hit in another project
Added breakpoints filter for source types existing multiple times in the
workspace and not reachable from the current resource mapping.

To avoid possible performance issues a type name cache is added which
remembers if a type is defined multiple times in the workspace.

Note, that PDE like launch configurations which do not specify any
resources in the launch configuration and launch configurations with non
standard classpath (like scrapbook, see bug 484036) are not affected by
this change.

Change-Id: Ibd0e1109a7c4fbd9cde0bfa2719927e55bfcba0a
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>