Bug 491299 - Classpath tab for Java application has wrong colors
Non editable entries are white text on black background on Linux with
Gtk default theme Adwaita (due to usage of COLOR_INFO_* constants, which
are ment for popups) which is way too intrusive in the UI.
Additionally when using Eclipse dark theme having same colors makes it
undistinguishable from the rest of the UI using dark pallette.
As the bug is really similar to bug 488825 I'm taking the same approach
and will make the respective class not implement IColorProvider.

Change-Id: I5014e893e7c322d19c76b47c474fa437132ce820
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>