Bug 491302 - Java/Debug/Logical Structure pref page has wrong colors
Non editable entries are white text on black background on Linux with
Gtk default theme Adwaita (due to usage of COLOR_INFO_* constants, which
are ment for popups) which is way too intrusive in the UI. 
Additionally when using Eclipse dark theme having same colors makes it
undistinguishable from the rest of the UI using dark pallette. Make the
class not implement IColorProvider as this should be handled by the
theme. 
Remove useless non-javadoc comments and redundant types while at it.

Change-Id: I5ef176e329e2ace85453fd682431ea67f6713c0f
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>