Bug 516278 - Lambda variables (closures) are not recognized by debugger variable inspector
Variables from enclosing method should be now resolved inside
lambdas in many simple cases.

The two corner cases left are:
1) The variables from enclosing lambda bodies (lambdas nested in
lambdas), because there is no easy way to identify the proper stack
frame for the parent lambda expression.
2) The "outer" variables if the lambda was not directly used in the
defining method but passed by as argument to another method (like
listeners etc).

In all cases the variables are attached to the stack below the current
but they are unnamed (arg$1, arg$2, ordered by they occurrence in the
source code).

A generic solution which will cover all use cases should parse the
bytecode or sourcecode of the lambda body (depending on the enclosing
method type) to get the current variable index in the lambda body and
access the variable from the previous stack with this index.

Change-Id: Iefbade3d8711deddf5a951666f2455e770c4c5cc
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>