Bug 452665 - [jdt] Socket operation on nonsocket: configureBlocking
Make sure we don't initialize server socket with default port if we
don't get the address. This seem to cause "Socket operation on
nonsocket: configureBlocking" error later.

Still unclear how the null address could be passed to the service except
we are called from startListening() which according to the comment in
the code is not supported by JDT.

Change-Id: Ib4609212a0f06997fa7be4b7cd348c3b578ee699
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>