Bug 525971 - Run to line in lambdas fail
RunToLineAdapter uses JLS4 thus it can't parse the lambda lines and
considers the cursor line as not a valid code while it is.
Move to JLS9 so it can parse any java construct fixes the issue.
Removed the non-javadoc lines to not clutter the codebase.

Change-Id: I58dcf86f101a63fc46dd90e61dcf1a570bd5c185
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>
(cherry picked from commit 600c6b52bb01fcfa7cabf16261937642dfff5a74)