Bug 528808 - Support for Thread.name change in the debuggee JVM
Added automatic update for thread names in the Debug view to allow
developers see real values of thread names without closing/re-opening
the view. This allows debugger to show job names for worker threads as
requested in bug 223492.

Debugger installs (user invisible) method entry breakpoint on
Thread.setName(String) method to be informed if the thread is changing
its name. On a breakpoint hit, debug STATE CHANGE event with thread as
source is sent to listeners, and so Debug view can update the thread
representation.

This is controlled by a new debugger option "Java -> Debug -> Listen to
thread name changes", which is set to "true" by default.

Change-Id: I0ccd9040c269d2be33c48b8d45763633ebf30e33
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>