Bug 536811 - ThreadNameChangeTests failed with No value present
Its possible that ThreadNameChangeTests attempts to retrieve a thread
necessary for the test, before JDIDebugTarget.ThreadStartHandler has
added this thread to the set of threads in JDIDebugTarget.

This change adds a timeout of five seconds to the test, to allow
processing in ThreadStartHandler to finish in a reasonable time span.

Change-Id: Ia481c77932536379c70196e450cd7d59e71c1195
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>