Bug 536140 - Exception logged in Error Log view when terminating debug
Since termination of a debug target can occur at any point, its possible
for code to check if the JVM is terminated and nevertheless later on run
into a terminated state. An example of this is found in
JavaBreakpoint.createRequests, which can throw an exception despite
checking if the JVM is available.

Such exceptions are then logged later on by JDIDebugElement.logError,
since they don't contain a VMDisconnectedException as cause. This change
ensures that cases in which the debuggee JVM disconnect was detected
also supply the VMDisconnectedException as cause. This avoids unwanted
error log entries.

This change also adds a few other, obvious, exception causes.

Change-Id: Ic9b981ff556bf3ce6ae728d587925e83c1f2b248
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>