Bug 537172 - NPE logged during testSuspendLongRunningCondition
During ConditionalBreakpointsTests.testEvalConditionOnStep a conditional
breakpoint is set in a loop. As the test stands, its possible that the
handling for this breakpoint coincides with the termination done at the
end of the test. If this occurs, a NPE will be thrown during the
condition evaluation. This is the case since the interpreter will ask
the terminated VM for a new value, which results in a null. Once the NPE
is thrown,
ConditionalBreakpointHandler.EvaluationListener.evaluationComplete is no
longer called. This leaves a lock which prevents conditional breakpoint
handling during the rest of the tests. This in turn causes numerous jobs
to hang.

This change ensures the conditional breakpoint cannot be hit during
termination. This is done by waiting for conditional breakpoint
evaluation to begin, before interrupting the evaluation as previously
done in the test. Additionally, the test snippet is adjusted to not run
a loop which can hit the conditional breakpoint and the changes for bug
516348 and bug 519382 are removed (since they are now obsolete).

Change-Id: I06729d18a9d5e9fbe0c0f58e84d8fe0309d90d92
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>