Bug 536943 - IOException below OutputStreamMonitor.read
To retrieve JVM info, StandardVMType.getLibraryInfo runs LibraryDetector
in an external process. A rare IOException is observed from one of the
output monitors of the process. The IOException indicates a closed
stream.

This change adds extra logging, in case the process which runs
LibraryDetector fails. Furthermore, a 2nd try is added in case
retrieving the library info failed.

Change-Id: If76b1dd3c3942ed4f8c4919bc299ae70f4319116
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>