Bug 540132 - ResourceException when updating breakpoint messages
Breakpoint messages are updated on breakpoint notifications. Those
usually are triggered in a job. Its therefore possible that the
respective project was deleted before such a job manages to update
breakpoint messages. In this case, a ResourceException is thrown.

With this change, a job is scheduled to update breakpoint messages. The
job will use a workspace modification lock and check whether markers
still exist, before updating their messages. This prevents the
ResourceException.

Change-Id: Ie57731dbca587bbe8c2b0afa8b77fe1507ef8e70
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>