Bug 544578 - No JRE sources found with Java 8 compliance Java 11 runtime
The fix for bug 525840 introduces a regression for the case of debugging
a Java project that has Java 8 on build path but runs with Java 11 as
runtime: The JRE sources cannot be found when stepping through the code.

This patch tries to find a source attachment for given runtime entry
also for the "modular" JRE case if no direct match was found by
inspecting all opened Java projects.

Change-Id: I72a5aa79c1d5b287c0f6321b88008e382f28895e
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>
Also-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>