Bug 546900 - Fix IO handling in JavaStacktraceConsole
If previous content of JavaStacktraceConsole could not loaded with a
single invocation of read the resulting content was corrupted and if eof
is reached before expected number of bytes are read it will block the
UI-thread for a long time.

Only on Windows: if console is empty on close stackTraceConsole.txt
should be deleted but failed and leave an empty file.

Change-Id: Id4007602d782fb7cbeaf486d8328149d09f5fd6f
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>