Bug 551029 - don't try to access WorkbenchWindow.showHeapStatus from non UI thread
AbstractDebugTest.setPreferences() tried to access
WorkbenchWindow.showHeapStatus() to hide the status monitor (WHY???),
which is of course a very bad idea if Window needs access to SWT and we
are running in non UI thread.

This worked before as long as the heap status widget was not yet created
(because there was no widget to hide), but now we see that dong this was
not a good idea.

I don't think this monitor widget hiding is really needed for debug
tests, so simply remove the code doing this in the wrong thread.

Change-Id: Ib55e135d4e82615ed44613d4c98d714f7eb06747
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>