Bug 560392 - Evaluate Lambda Field variables
When passing method parameters to a returning lambda object from the
method the variable is embedded into lambda object as val$variable_name.
This naming convention cause the evaluator to skip this variable pushed
into evaluator engine and also to not matched with the generated
snippet.

This fix provide way to push this variable into evaluator engine and at
the same time the generated snippet is adjusted to the above mentioned
naming convention,

Change-Id: I3338ab467e107166690b06e05510981afe0b0fe3
Signed-off-by: gayanper <gayanper@gmail.com>