Bug564486 - Fix variable evaluation for anon classes and lambdas
The evaluation engine fails when the lambda is inside a anon class and
the evaluation is done inside the lambda. The root cause was the eval
method which is static throws a compilation problem due to it is created
in the anon class. Now the problem is ignored as other problems in
ASTEvaluationEngine

Change-Id: I287cf2d02561e49b90c32f7124c0e1e90d840279
Signed-off-by: gayanper <gayanper@gmail.com>