Bug 565982 - Hit count method breakpoint always hit without line info
If a Java project has the compiler options set to not produce line info,
method entry breakpoints with a hit count are hit always despite the hit
count.

This change adjusts JavaMethodBreakpoint.handleHitCount(LocatableEvent,
Integer) to use the same implementation as
JavaBreakpoint.expireHitCount(Event) when the breakpoint hit count is
reached. This results in the expected behaviour for method entry
breakpoints with a hit count: the breakpoint is hit upon reaching the
hit count and is then disabled until the Java application launch is
done, or the hit count changes, or the breakpoint is re-enabled (e.g. by
the user).

Change-Id: Ia35743f310b1b20d90290f40794bea2e9c37d39d