Bug 568813 - Empty inspect window and hover
The empty hover on BigSur is caused due to SWT Bug 567787- Setting Shell visible doesn't send SWT.SetData for Virtual Tree/Table. Due to this the virtual tree in the hover is not populated when set to visible.

The workaround is to call getText on the item so that it is already created when hover is shown.

Change-Id: I9d49902c4013f97db4005c948205aa37ccb4c327