Bug567801 : Fix generic signature resolution in evaluation
When resolving generic signatures of variables available for
 evaluation, it was not considered all types of generic signatures.
This cause some of the generic signatures to resolved into Object
type when they can be clearly resolved to the correct expected type.

The fix will consider all generic cases including bounded and
wild-card scenarios and resolve the correct types for variables.


Change-Id: Ic40cec57b81654eeb7024b5d8d27dd7a1957290f
Signed-off-by: Gayan Perera <gayanper@gmail.com>