Bug569600 : Fix debug inspection popup issues
This will mainly fix two issues, first is the issue due to wrong focus
hand over events on some platforms which cause the inspection popup to
close it self as soon as it is opened. The second issue which is AIOBE
is due to trying to evaluate a empty snippet string.

Change-Id: I9bfeceff94b0469d54b98839af2b0f80009e56a8
Signed-off-by: Gayan Perera <gayanper@gmail.com>