Bug 572782 - Fix evaluation errors due to recursive type bounds
When where are recursive type bounds which contains the type declaration
name which the current type bounds are defined for, the source generator
add these bounds into generated run method which cause the compilation
issue. The fix avoid adding such type bounds since they are not valid
for a method declaration.

Change-Id: Ia8fe4d2bad3ec66366964931f03faa65897243d2
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/179463
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Sarika Sinha <sarika.sinha@in.ibm.com>