Bug 572629 - Add support debug hovers on chain of objects
Extend the current debug hovering for chain objects by using performing
a code evaluation for such expressions. Before this change the hovering
only try to reads variables that are available in current frame and
member variables. All other chained variables which are identified as
qualified names are evaluated using ASTEvaluationEngine, and for
synthetic members such as array.length we try to select the node of the
array field by moving the offset and evaluate it in the same way as
qualified names.


Change-Id: Ida6f14f85b5aa3fcc6019216318a03ba694c19a0
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/179566
Tested-by: Sarika Sinha <sarika.sinha@in.ibm.com>
Reviewed-by: Sarika Sinha <sarika.sinha@in.ibm.com>