Bug 572782 - Fix evaluation errors due to recursive type bounds
When where are recursive type bounds which contains the type declaration
name which the current type bounds are defined for, the source generator
add these bounds into generated run method which cause the compilation
issue.

The fix try take the approach of adding the type parameters which are
only relevant, that is if the current evaluation is in static context
we only add the method type parameters which is under evaluation and if
we are inside none static context we don't add type parameters since
they are visible for the generated method.


Change-Id: I68303f14308e5922d9236b8abc1c3d774a52ba5a
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/179692
Tested-by: Sarika Sinha <sarika.sinha@in.ibm.com>
Reviewed-by: Sarika Sinha <sarika.sinha@in.ibm.com>
Reviewed-by: Jesper Moller <jesper@selskabet.org>