Bug 573479 - skip already visited entries when expanding classpath
DefaultProjectClasspathEntry.expandProject() performs a depth first
iteration over the projects classpath. It tries to prune already visited
sub trees by checking if the entry is already in the expandedPath list.
One code path however transforms the entry before adding it to the list.
In that case the pruning does not work.
Fixed by introducing a dedicated set to keep track of already visited
entries.

Change-Id: I0c0946b597e225645501a410a9f14614fa6cae62
Signed-off-by: Andreas Huber <ahe@lucares.de>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/180466
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>