Bug 547462 - Fix inconsistency in infix expression evaluation
When evaluating a infix expression which contains extended operands,
then the number of no-op that are pushed are not stored in instructions
when leaving the infix expression. This cause problem with instruction
order when in situations where their is a prefix expression, where the
prefix expression will be popped too late into instructions.

The fix solves this issue by storing/popping the no-op from stack equal
to the number of extended operands in the infix expression before it
leaves.

Change-Id: I5451b68e34a00007736ec24ff421cd112b6f85fd
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/180551
Reviewed-by: Jesper Moller <jesper@selskabet.org>
Tested-by: JDT Bot <jdt-bot@eclipse.org>