Bug 575039 - Fix evaluations inside JVM classes under java9 and above
When compiling the evaluation snippet with java runtime classes, the
compiler report issues due to java9 module system restrictions. Since
the snippet is considered to be on a unnamed module, it introduce
conflicts, visibility issues etc.

The solution tries to use the existing internal compiler flag to switch
the compiler into debug friendly mode which will ignore the module
system issues when compiling the snippet. This is only done if the
runtime vm is above java 1.8 only.

Change-Id: I8e83f045517feb7cfa8faf2d2dbb6a72b2774d00
Signed-off-by: Gayan Perera <gayanper@gmail.com>
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/183703
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>