Bug 576999: Wait for selectFrame to complete before resuming test
Doing a select frame causes (after some delay) a selectAndReveal on the
editor for the frame's location. Make sure this selectAndReveal is
complete before continuing the test.

Prior to this patch, sometimes the frame's selectAndReveal ran after the
key one in the test and the wrong thing would be selected in the editor
for the test.

triggerWinGerrit

Change-Id: I7fbf951886ff157baa0b30fe5e3c73fa2aae0368
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/191738
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Tested-by: Releng Bot <releng-bot@eclipse.org>
Reviewed-by: Jonah Graham <jonah@kichwacoders.com>
Reviewed-by: Sarika Sinha <sarika.sinha@in.ibm.com>