Bug 579439 - adapt LineTrackerTests.testFlood
The tests logs more then the default buffer size can keep. And logs
faster then the console can draw. With the changes of bug 575275 the
console is only formated for the visible portion of the console buffer.

see
org.eclipse.ui.internal.console.IOConsolePartitioner.checkBufferSize()

Fix for the test: unlimit the console buffer during this test.

Change-Id: I69803f63816469570bab9d49c4ddf2c33875a44a
Reviewed-on: https://git.eclipse.org/r/c/jdt/eclipse.jdt.debug/+/192281
Tested-by: JDT Bot <jdt-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>