Remove leftover from bug 148006 fix #2
Bug 148006 introduced a coding mistake.

It makes no sense to join on the job instance. What was meant, was the
outer JavaStackTraceConsole instance. So we still see errors if console
changes the document on which the MatchJob is working.

But fixing that would introduce deadlocks like in bug 451797

So simply delete this no-op line that was there since ever.

Note: the bug that was meant to be fixed: when the console format runs
in parallel to the MatchJob running, it should wait.

But waiting in UI thread mean *freeze*.