Bug 575131 - Deadlock with conditional breakpoints that suspend VM
Adjusts JDIDebugTarget resumeThreads() and
suspendThreads() to not resume resp. suspend threads that are handling
breakpoints.

Also adds a test to reproduce the deadlock, as well as a stress test
suite that will repeat the new test (since the deadlock is sporadic).
The stress test suite is not executed during automated tests.

Change-Id: Icfcd5ebb04e77e437bc308d5cb6911a77eb3b673
Signed-off-by: Simeon Andreev <simeon.danailov.andreev@gmail.com>