Run StringSubstitutionTests with Project Explorer
Running with Navigator is pointless test as this view is not visible by
default.
Changed the test to not use .classpath file but rather existing java
file from the project in order to not change the default filters for the
view but this doesn't affect the test as it's about variables resolving.
Removed useless non-javadoc comments.