Evaluate "without_test_code" when resolving the runtime classpath #327 (#328)
When resolving the runtime classpath for project dependencies the
attribute "without_test_code" has to be evaluated so that paths marked
as test code from the included project do not contribute to the
classpath. The contained classes or resources should not be accessible
(e.g. via reflection).

Fixes https://github.com/eclipse-jdt/eclipse.jdt.debug/issues/327