Make the snippetsupport.jar as part of the maven build
Currently snippetsupport.jar must be build in the ide using s ant
script.

This changes the build to use maven only rebuilding the jar on each
compile so class differences can be discovered automatically. The jar
still needs to be checked in as it is used internally  as a resource.