Do not automatically detect JVMs on CI systems
Automatic detection on CI systems can cause a lot of trouble (e.g.
memory consumption, time to read the jvms, old jvms lingering around)
and is generally not useful as a CI has usually a dedicated and fixed
setup to run with.

This adds a check for the common 'CI' environment variable that is
usually defined by all usual CI/CD pipelines to indicate a job is
running on a CI server to skip the detection regardless of preferences
configuration. If one absolutely want the feature even for CI, the
system-property 'DetectVMInstallationsJob.disabled' can be set to false
to restore previous behavior.