Allow JavaVariableLabelProvider's presentation to be overridden
This mostly reverts a4eae7ff2f3dc3f51234053cc4a9374c2f6f8b51 (#580) for this
file, as well as some code cleanup. Instead of making the fLabelProvider
final, we initialize it to a default value and provide getters/setters.

I don't understand this code enough to know why fLabelProvider is
static, given that label providers are not meant to be shared, so I left
it alone. Otherwise it would make more sense to have it as an instance
field, maybe defined in an overloaded constructor.