Don't use Display.syncExec() for refresh in RuntimeClasspathViewer
The listener may be called from a background thread that already holds
some locks. To refresh RuntimeClasspathViewer UI in the current case,
syncExec is not needed, asyncExec is enough and it would not acquire the
UI lock from an unknown thread.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.debug/issues/672