Avoid waiting for reconciler job
After https://github.com/eclipse-platform/eclipse.platform.ui/pull/3025
every opened editor means there will be a reconciler job running, which
is not what we do NOT want to wait for. AbstractReconciler.class is the
family object we can check for, so add that family to the list of
excluded jobs we don't care.

Added getUsualJobFamiliesToIgnore() and changed waitForJobs() to use
that by default if nothing else is given.

Fixes https://github.com/eclipse-jdt/eclipse.jdt.debug/issues/721