Make determination of execution environment compliance generic
Currently one needs to add a new if for each new java version while the
value is only matched against the number in the string.

This now uses a pattern and check the result is larger than Java 1.8
version and then returns the number as a string so it automatically
adapt to new java versions.