Select the best matching JRE for an EE from the list of compatible JVMs (#761)
Currently the selection of a compatible JRE for an EE seems quite random
for the user, e.g. it can happen that for an EE10 with some JRE11/17/24
it chooses 24. This is actually a bad choice as newer java might
_remove_ API during release, so the _lowest_ possible java version among
others would be the best choice here.

This now do the following:

- introduce a new method IExecutionEnvironment#getCompatibleVM that
selects the best matching from the set of selected JVMs (but ignoring
the default)
- in the JREContainerInitializer check if there is a default vm
selected, then use that
- if not ask the EE for a compatible one (with the new method)